refactor: move input to inputfield (#196)
