+++ src/InputField/index.js
@@ -2,7 +2,7 @@ import propTypes from 'prop-types'
 import React, { Fragment } from 'react'
 import css from 'styled-jsx/css'
 
-import { Input } from '../Input'
+import { Input } from './Input.js'
 import { LabelFilled, LabelOutlined } from '../FieldLabel'
 import { theme } from '../theme'
